+++ src/tooling/piral-cli/src/rules/pilet-has-non-conflicting-css.ts
@@ -1,0 +1,13 @@
+import { isMonorepoPackageRef, findLatestVersion } from '../common';
+import { PiletRuleContext } from '../types';
+
+export type Options = 'suggest' | 'required' | 'ignore';
+
+/**
+ * Checks if the pilet has some CSS rules that might conflict with other pilets.
+ */
+export default async function (context: PiletRuleContext, options: Options = 'suggest') {
+  if (options !== 'ignore') {
+    //TODO
+  }
+}